moviedb-318 Display matching movies from IMDB in EditMovieGUI window.
Directories and fielnames for gui tests reorganized. Part II.
