+++ test/gui/test_neurons.py
@@ -1,7 +1,7 @@
 """Test module."""
 
 #  Copyright (c) 2020-2024. Stephen Rigden.
-#  Last modified 2/16/24, 9:16 AM by stephen.
+#  Last modified 3/9/24, 10:00 AM by stephen.
 #  This program is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
