+++ test/gui/test_tk_facade.py
@@ -1,7 +1,7 @@
 """ Test module"""
 
 #  Copyright (c) 2024-2024. Stephen Rigden.
-#  Last modified 3/9/24, 9:39 AM by stephen.
+#  Last modified 3/9/24, 10:00 AM by stephen.
 #  This program is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
